# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CampusCoin API
  version: 0.1.0
  description: CampusCoin ("Smart Spending, Student Style") REST API. Money amounts are decimal strings (never floats); errors follow RFC 9457 `application/problem+json`; every route scoped to the caller's own data returns 404 (not 403) for another user's resource.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token (EdDSA JWT) from `POST /auth/login`, sent as `Authorization: Bearer <token>`."
  schemas:
...
    updateImportRowsSchema:
      type: object
      properties:
        options:
          type: object
          properties:
            dateFormat:
              type: string
              enum:
                - YYYY-MM-DD
                - DD/MM/YYYY
                - MM/DD/YYYY
            mapping:
              type: object
              properties:
                date:
                  type:
                    - integer
                    - "null"
                  minimum: 0
                  maximum: 49
                amount:
                  type:
                    - integer
                    - "null"
                  minimum: 0
                  maximum: 49
                type:
                  type:
                    - integer
                    - "null"
                  minimum: 0
                  maximum: 49
                description:
                  type:
                    - integer
                    - "null"
                  minimum: 0
                  maximum: 49
                category:
                  type:
                    - integer
                    - "null"
                  minimum: 0
                  maximum: 49
              required:
                - date
                - amount
                - type
                - description
                - category
              additionalProperties: false
          additionalProperties: false
        rows:
          type: array
          items:
            type: object
            properties:
              rowNumber:
                type: integer
                minimum: 2
              categoryId:
                type: integer
                exclusiveMinimum: 0
              selected:
                type: boolean
            required:
              - rowNumber
            additionalProperties: false
          maxItems: 200
        setAllSelected:
          type: boolean
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: Machine-readable error slug, e.g. "validation-failed".
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required:
              - field
              - message
          description: Present on 422 validation-failed responses, one entry per invalid field.
      required:
        - type
        - title
        - status
      description: RFC 9457 application/problem+json body.
...
paths:
...
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: List notifications.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            enum:
              - "true"
              - "false"
            default: "false"
          required: false
          name: unreadOnly
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/notifications/read-all:
    post:
      tags:
        - Notifications
      summary: Mark every notification read.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/notifications/stream-ticket:
    post:
      tags:
        - Notifications
      summary: Get a one-time ticket for the SSE stream.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/notifications/{id}/read:
    post:
      tags:
        - Notifications
      summary: Mark one notification read.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,20}$
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/notifications/stream:
    get:
      tags:
        - Notifications
      summary: Server-Sent Events stream of live notifications (auth via `?ticket=`, not Bearer).
      security: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /api/v1/imports/template:
    get:
      tags:
        - Imports
      summary: Download the CSV import template.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/imports:
    post:
      tags:
        - Imports
      summary: Upload a CSV file to start an import batch (multipart/form-data, 10/hour).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "413":
          description: Upload exceeds the size limit.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "415":
          description: Unsupported file type.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/imports/{id}:
    get:
      tags:
        - Imports
      summary: Get an import batch (preview rows, paginated).
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
        - schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: string
            enum:
              - all
              - errors
              - duplicates
              - needs_review
            default: all
          required: false
          name: filter
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - Imports
      summary: Discard an import batch.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/imports/{id}/rows:
    patch:
      tags:
        - Imports
      summary: Edit preview rows before commit.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateImportRowsSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/imports/{id}/commit:
    post:
      tags:
        - Imports
      summary: Commit an import batch (creates transactions). Supports `Idempotency-Key`.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Conflict (duplicate, already committed, optimistic-lock mismatch, etc).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/imports/{id}/errors.csv:
    get:
      tags:
        - Imports
      summary: Download the row-level error report as CSV.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...